[BFCL] Refactor Model Handler into OSS and Proprietary Components (#612)
This PR reorganizes the model handler by splitting it into two distinct
components: an Open Source (OSS) model handler and a Proprietary model
handler. This change is part of a series of updates that address the
tasks outlined in issue #510.

---------
