+++ berkeley-function-call-leaderboard/.env.example
@@ -1,0 +1,20 @@
+# [OPTIONAL] Required for LLM generation step
+# Provide the API key for the model(s) you intend to use
+OPENAI_API_KEY=sk-XXXXXX
+MISTRAL_API_KEY=
+FIREWORKS_API_KEY=
+ANTHROPIC_API_KEY=
+NVIDIA_API_KEY=nvapi-XXXXXX
+YI_API_KEY=
+
+COHERE_API_KEY=
+USE_COHERE_OPTIMIZATION=False # True/False
+
+DATABRICKS_API_KEY=
+DATABRICKS_AZURE_ENDPOINT_URL=
+
+# [OPTIONAL] Required for evaluation of `exec` test group 
+RAPID_API_KEY=
+EXCHANGERATE_API_KEY=
+OMDB_API_KEY=
+GEOCODE_API_KEY=
