+++ berkeley-function-call-leaderboard/bfcl/eval_checker/eval_runner.py
@@ -4,6 +4,7 @@ from eval_runner_helper import *
 from eval_checker_constant import TEST_COLLECTION_MAPPING
 from tqdm import tqdm
 import argparse
+from dotenv import load_dotenv
 
 
 # NOTE: This file should be run in the `eval_checker` directory
@@ -487,5 +488,8 @@ if __name__ == "__main__":
             # This is differnet than the model name format that the generation script "openfunctions_evaluation.py" takes in (where the name contains "/").
             # We patch it here to avoid confusing the user.
             model_names.append(model_name.replace("/", "_"))
+
+    load_dotenv(dotenv_path="../../.env", verbose=True, override=True)  # Load the .env file
+    assert os.getenv("USE_COHERE_OPTIMIZATION") in ["True", "False"], "USE_COHERE_OPTIMIZATION in .env must be set to True or False"
 
     runner(model_names, test_categories, api_sanity_check)
