+++ berkeley-function-call-leaderboard/bfcl/model_handler/oss_model/salesforce.py
@@ -1,6 +1,6 @@
 import json
 
-from bfcl.model_handler.oss_handler import OSSHandler
+from bfcl.model_handler.oss_model.base_oss_handler import OSSHandler
 from bfcl.model_handler.model_style import ModelStyle
 
 SYSTEM_PROMPT = """You are an AI assistant for function calling. 
@@ -30,7 +30,7 @@ The example format is as follows. Please make sure the parameter type is correct
 """
 
 
-class xLAMHandler(OSSHandler):
+class SalesforceHandler(OSSHandler):
     def __init__(
         self, model_name, temperature=0.001, top_p=1, max_tokens=512, dtype="bfloat16"
     ) -> None:
