+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/databricks.py
@@ -1,31 +1,34 @@
-import time, os
-from openai import OpenAI
-from bfcl.model_handler.handler import BaseHandler
+from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.model_style import ModelStyle
-from bfcl.model_handler.utils import ast_parse
 from bfcl.model_handler.utils import (
     func_doc_language_specific_pre_processing,
     system_prompt_pre_processing_chat_model,
     combine_consecutive_user_prompr,
+    ast_parse,
 )
 from bfcl.model_handler.constant import (
     DEFAULT_SYSTEM_PROMPT,
 )
+import time, os
+from openai import OpenAI
+import re
 
 
-class NvidiaHandler(BaseHandler):
+class DatabricksHandler(BaseHandler):
     def __init__(self, model_name, temperature=0.001, top_p=1, max_tokens=1000) -> None:
         super().__init__(model_name, temperature, top_p, max_tokens)
         self.model_style = ModelStyle.OpenAI
+
+        # NOTE: To run the Databricks model, you need to provide your own Databricks API key and your own Azure endpoint URL.
         self.client = OpenAI(
-            base_url="https://integrate.api.nvidia.com/v1",
-            api_key=os.getenv("NVIDIA_API_KEY"),
+            api_key=os.getenv("DATABRICKS_API_KEY"),
+            base_url=os.getenv("DATABRICKS_AZURE_ENDPOINT_URL"),
         )
 
     def inference(self, prompt, functions, test_category):
         functions = func_doc_language_specific_pre_processing(functions, test_category)
+
         prompt = system_prompt_pre_processing_chat_model(prompt, DEFAULT_SYSTEM_PROMPT, functions)
-        
         prompt = combine_consecutive_user_prompr(prompt)
         message = prompt
 
@@ -37,46 +40,44 @@ class NvidiaHandler(BaseHandler):
             max_tokens=self.max_tokens,
             top_p=self.top_p,
         )
+
         latency = time.time() - start_time
         result = response.choices[0].message.content
-        input_token = response.usage.prompt_tokens
-        output_token = response.usage.completion_tokens
-        metadata = {
-            "input_token_count": input_token,
-            "output_token_count": output_token,
-            "latency": latency,
-            "processed_message": message,
-        }
+        metadata = {}
+        metadata["input_token_count"] = response.usage.prompt_tokens
+        metadata["output_token_count"] = response.usage.completion_tokens
+        metadata["latency"] = latency
+        metadata["processed_message"] = message
         return result, metadata
 
     def decode_ast(self, result, language="Python"):
-        result = result.replace("\n", "")
-        if not result.startswith("["):
-            result = "[" + result
-        if not result.endswith("]"):
-            result = result + "]"
-        if result.startswith("['"):
-            result = result.replace("['", "[")
-            result = result.replace("', '", ", ")
-            result = result.replace("','", ", ")
-        if result.endswith("']"):
-            result = result.replace("']", "]")
-        decode_output = ast_parse(result, language)
+        func = re.sub(r"'([^']*)'", r"\1", result)
+        func = func.replace("\n    ", "")
+        if not func.startswith("["):
+            func = "[" + func
+        if not func.endswith("]"):
+            func = func + "]"
+        if func.startswith("['"):
+            func = func.replace("['", "[")
+        try:
+            decode_output = ast_parse(func, language)
+        except:
+            decode_output = ast_parse(result, language)
         return decode_output
 
     def decode_execute(self, result, language="Python"):
-        result = result.replace("\n", "")
-        if not result.startswith("["):
-            result = "[" + result
-        if not result.endswith("]"):
-            result = result + "]"
-        if result.startswith("['"):
-            result = result.replace("['", "[")
-            result = result.replace("', '", ", ")
-            result = result.replace("','", ", ")
-        if result.endswith("']"):
-            result = result.replace("']", "]")
-        decode_output = ast_parse(result, language)
+        func = re.sub(r"'([^']*)'", r"\1", result)
+        func = func.replace("\n    ", "")
+        if not func.startswith("["):
+            func = "[" + func
+        if not func.endswith("]"):
+            func = func + "]"
+        if func.startswith("['"):
+            func = func.replace("['", "[")
+        try:
+            decode_output = ast_parse(func, language)
+        except:
+            decode_output = ast_parse(result, language)
         execution_list = []
         for function_call in decode_output:
             for key, value in function_call.items():
