+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/nexus.py
@@ -1,5 +1,5 @@
 from bfcl.model_handler.model_style import ModelStyle
-from bfcl.model_handler.handler import BaseHandler
+from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.utils import (
     ast_parse,
     func_doc_language_specific_pre_processing,
